KS-1-4 change format

+++ src/main/kotlin/lesson_1/lesson_1_task_4.kt
@@ -3,7 +3,7 @@ package org.example.lesson_1
 fun main() {
 
     val trajectoryLenght: Long = 40868600000L
-    val crewMemberAge: Short = 27
+    val crewMemberAge: Byte = 27
     val partOfDay: Float = 0.075f
     val seconds: Short = 6480
     val partOfYear: Double =  2.0547945205479453E-4
